lesson_10/requirements.txt
apache-airflow[gcp]==2.5.1
google-cloud-storage==1.42.3
